\name{frmhet-package}
\alias{frmhet-package}
\docType{package}
\title{Regression Analysis of Fractional Responses Under Unobserved Heterogeneity}
\description{Estimation and specification analysis of fractional regression models with neglected heterogeneity and/or endogenous covariates.
}
\details{
\tabular{ll}{
Package: \tab frm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-18\cr
License: \tab GPL-2\cr
}
}
\author{Joaquim J.S. Ramalho <jsr@uevora.pt>}
\references{Ramalho, E.A. and J.J.S. Ramalho (2015),
"Moment-based estimation of nonlinear regression models with boundary outcomes and endogeneity, with applications to nonnegative and fractional responses", \emph{Econometric Reviews}, forthcoming (DOI: 10.1080/07474938.2014.976531).
}